Remove cypress-keycloak-command
Reason: Don't want to have actual tokens stored in the codebase
