--- cypress/fixtures/users/fake-user.json
@@ -1,15 +1,0 @@
-{
-  "fakeLogin": {
-    "access_token": "eyJhbGciOiJSUzI1NiIsInR5cCIgOiAiSldUIiwia2lkIiA6ICJQZUlFYlpQUm5xaTk2ZDl6RFJZT3B5RFBqVHdkei1tcF9kRHJPYmJ4Uko0In0.eyJleHAiOjE2MDE0ODU2NTYsImlhdCI6MTYwMTQ4NTM1NiwiYXV0aF90aW1lIjoxNjAxNDg1MzM2LCJqdGkiOiI0MzkxZmEyOC1jY2FiLTRmOTgtOGM2Yy00ODg0OWQwMDQzM2QiLCJpc3MiOiJodHRwczovL3Nzby1kZXYucGF0aGZpbmRlci5nb3YuYmMuY2EvYXV0aC9yZWFsbXMvOHdsNng0Y3AiLCJzdWIiOiJhYjQyZjQ0My0yNjI0LTRlMjItODIzZi1jMTUxYTJhMmIzZWIiLCJ0eXAiOiJCZWFyZXIiLCJhenAiOiJ3cHMtd2ViIiwibm9uY2UiOiI1OGQ4MmU0NS02ODlmLTQyZWQtODc2OS1jMGFlNmE5MWI3YzAiLCJzZXNzaW9uX3N0YXRlIjoiYjMxYTVhYjItY2RmMC00MzYxLTgwZjctOGNlYmRhM2I2MTY2IiwiYWNyIjoiMCIsImFsbG93ZWQtb3JpZ2lucyI6WyIqIl0sInNjb3BlIjoib3BlbmlkIHByb2ZpbGUgZW1haWwiLCJlbWFpbF92ZXJpZmllZCI6ZmFsc2UsIm5hbWUiOiJLeXViaW4gSGFuIiwicHJlZmVycmVkX3VzZXJuYW1lIjoia2hhbkBpZGlyIiwiZ2l2ZW5fbmFtZSI6Ikt5dWJpbiIsImZhbWlseV9uYW1lIjoiSGFuIiwiZW1haWwiOiJreXViaW4uaGFuQGdvdi5iYy5jYSJ9.NtCsBeBHsK2JvyEElhvxbmmy-wwWYsSoBMFTSKrQCcHvVSW_yb3-D-VwjK280xZDrpDNVixep3CDWr85d42z8I3JVxr-wih8pYu3TxqSkSUNcSN4sBQEkGekwY5DXIeHckm82n4oXTF20C9d0tHoPGnV9MnktQTqWmO5ZD10Ut9m1oc2-IJyfw95YCoc4H1hkmoKk8RL73UMIcgv1sl4iD0DdGJUR4mXJOnKQuN-XKmwl773xfHYFeODakrbVNnps1iUmxPT4MCfm97rHgG477t669s0MkBSD7I1hX_dy6jOZvNTA-_LybfGO-FVCxibit7PFGDFFDpqin7sQ4oF4g",
-    "refresh_token": "eyJhbGciOiJIUzI1NiIsInR5cCIgOiAiSldUIiwia2lkIiA6ICI2MTBjMDg2Ni01NWU2LTQyNzQtYjViOC1lMDgwMWE0YTNkYzAifQ.eyJleHAiOjE2MDE0ODcxNTYsImlhdCI6MTYwMTQ4NTM1NiwianRpIjoiYjFhMWQ5MjYtYzJkYS00MmI3LTgyN2QtODRiYjFhZDNkNDgyIiwiaXNzIjoiaHR0cHM6Ly9zc28tZGV2LnBhdGhmaW5kZXIuZ292LmJjLmNhL2F1dGgvcmVhbG1zLzh3bDZ4NGNwIiwiYXVkIjoiaHR0cHM6Ly9zc28tZGV2LnBhdGhmaW5kZXIuZ292LmJjLmNhL2F1dGgvcmVhbG1zLzh3bDZ4NGNwIiwic3ViIjoiYWI0MmY0NDMtMjYyNC00ZTIyLTgyM2YtYzE1MWEyYTJiM2ViIiwidHlwIjoiUmVmcmVzaCIsImF6cCI6Indwcy13ZWIiLCJub25jZSI6IjU4ZDgyZTQ1LTY4OWYtNDJlZC04NzY5LWMwYWU2YTkxYjdjMCIsInNlc3Npb25fc3RhdGUiOiJiMzFhNWFiMi1jZGYwLTQzNjEtODBmNy04Y2ViZGEzYjYxNjYiLCJzY29wZSI6Im9wZW5pZCBwcm9maWxlIGVtYWlsIn0.oEX0ZGbU96rRBstEiqHVJx7kHQknXEzLTc2uGOElOHw",
-    "id_token": "eyJhbGciOiJSUzI1NiIsInR5cCIgOiAiSldUIiwia2lkIiA6ICJQZUlFYlpQUm5xaTk2ZDl6RFJZT3B5RFBqVHdkei1tcF9kRHJPYmJ4Uko0In0.eyJleHAiOjE2MDE0ODU2NTYsImlhdCI6MTYwMTQ4NTM1NiwiYXV0aF90aW1lIjoxNjAxNDg1MzM2LCJqdGkiOiI1ZjBmODA2ZC1hMGVlLTQ2M2MtOTJiMy1lZDI2NWNhYmNjMmQiLCJpc3MiOiJodHRwczovL3Nzby1kZXYucGF0aGZpbmRlci5nb3YuYmMuY2EvYXV0aC9yZWFsbXMvOHdsNng0Y3AiLCJhdWQiOiJ3cHMtd2ViIiwic3ViIjoiYWI0MmY0NDMtMjYyNC00ZTIyLTgyM2YtYzE1MWEyYTJiM2ViIiwidHlwIjoiSUQiLCJhenAiOiJ3cHMtd2ViIiwibm9uY2UiOiI1OGQ4MmU0NS02ODlmLTQyZWQtODc2OS1jMGFlNmE5MWI3YzAiLCJzZXNzaW9uX3N0YXRlIjoiYjMxYTVhYjItY2RmMC00MzYxLTgwZjctOGNlYmRhM2I2MTY2IiwiYWNyIjoiMCIsImVtYWlsX3ZlcmlmaWVkIjpmYWxzZSwibmFtZSI6Ikt5dWJpbiBIYW4iLCJwcmVmZXJyZWRfdXNlcm5hbWUiOiJraGFuQGlkaXIiLCJnaXZlbl9uYW1lIjoiS3l1YmluIiwiZmFtaWx5X25hbWUiOiJIYW4iLCJlbWFpbCI6Imt5dWJpbi5oYW5AZ292LmJjLmNhIn0.jyRYpRCgvG4JAkcNpW7snCKq-RHoNwEx58_Y2MrwIvQgFYZAVC4MRDUopYOYrFp1nQztonnb4gCKY7_nYNPyttHkwZCLSjbNQieKPavn_X56BtXcvHayWa5KIMh1myMl1y2ICTas2qZT9z5806RzmJQ8cQAET73kzR5192iE6yKt_eqK_fw4CdAII1ZRPyj7Xwd7CoZnI7ykc136EhvXWCdnYWudui0jEJDpxNN-6H5TTOZkDE_WQTduuToKOr5vqlFn4evOOz4nMroLMKAZInnJSrSEZjHxykR0IXKXEzdOXp3jAvYrTzeUJ1cLyewNEh4RWRlzrtLaIY75P3S4oA",
-    "account": {
-      "username": "user",
-      "firstName": "Sample",
-      "lastName": "User",
-      "email": "sample-user@example",
-      "emailVerified": false,
-      "attributes": {}
-    }
-  }
-}
