+++ package.json
@@ -82,7 +82,6 @@
     "@typescript-eslint/parser": "^3.10.1",
     "axios-mock-adapter": "^1.17.0",
     "cypress": "^5.2.0",
-    "cypress-keycloak-commands": "^1.2.0",
     "eslint-config-prettier": "^6.10.0",
     "eslint-config-react": "^1.1.7",
     "eslint-plugin-prettier": "^3.1.2",
